Added some helper library modules
